% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRainbows.R
\name{aggregateData}
\alias{aggregateData}
\title{aggregate rayData}
\usage{
aggregateData(rayData, aggregateRainbows = FALSE,
  aggregateLambda = FALSE, nBreaks = 200)
}
\arguments{
\item{rayData}{data.table with rayData}

\item{aggregateRainbows}{aggregate over all rainbows (and set rainbowNo = 0). Default = FALSE}

\item{aggregateLambda}{aggregate over all lambdas (and set lambda = 0). Default = FALSE}

\item{nBreaks}{number of bins to aggregate into}

\item{cutoff}{exclude rows which are smaller than max*cutoff (0 = no cutoff)}
}
\value{
data.table with pdf of angular difference (angDDeg)
}
\description{
Aggregate rayData to pdf
}
\details{
Aggregates rayData over a range of angles. The range is implicitly
set by the number of bins. Cutoff removes values smaller than
a portion of max vaule
}
\author{
Fredrik Wartenberg
}
