% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{defaultParameters}
\alias{defaultParameters}
\title{Retrieve default paramenters}
\usage{
defaultParameters()
}
\value{
list with default parameters
}
\description{
Retrieve default paramenters
}
\author{
Fredrik Wartenberg
}
