% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{drop}
\alias{drop}
\title{water drop}
\usage{
drop(
  x = 0,
  y = 0,
  z = 0,
  R = 100,
  color = "green",
  alpha = "0.1",
  n = 1.33,
  ri = refractiveIndex
)
}
\arguments{
\item{x}{x-coordinate of centre}

\item{y}{y-coordinate of centre}

\item{z}{z-coordinate of centre}

\item{R}{radius}

\item{color}{rendering color}

\item{alpha}{rendring alpha}

\item{n}{refractive index (default)}

\item{ri}{function returning refractive index}
}
\value{
drop objecr
}
\description{
The water drop through which rays are traced
}
\details{
Contains both physical and graphical parameters for rendering.
}
\author{
Fredrik Wartenberg
}
