% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRainbows.R
\name{generateDataMatrix}
\alias{generateDataMatrix}
\title{Generate Rainbow Spectral Distributions}
\usage{
generateDataMatrix(universe, resolution = 10000, nColors = 1,
  rainbows = c(1, 2))
}
\arguments{
\item{universe}{The universe (typically one drop)}

\item{rainbows}{Rainbows to generate, e.g. c(1,2,4) for first, second and fourth rainbow}

\item{angularSteps}{Angular resolution for simulation steps}

\item{lambdaSteps}{Stepsize for lambda between visibleMin and visibleMax}

\item{intensity}{How to calculate the intensity in the spectrum:
count: just count the number of rays
fresnel: apply fresenel formulas to transmission and reflectio}
}
\value{
a data.table with the spectral and angular intesities for each rainbow
}
\description{
Generates the table with spectral and angular distribution of the different rainbows.
}
\details{
Central simulation function
}
\author{
Fredrik Wartenberg
}
