% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{generateRainbows}
\alias{generateRainbows}
\title{Generate Rainbows}
\usage{
generateRainbows(rainbows = c(1, 2), nColors = 5, resolution = 1, plot = TRUE)
}
\arguments{
\item{rainbows}{The raibows to be generated. Specified by a vector with the rainbow number.
Example: c(1,2,3) will generate the first three rainbows. Default = c(1,2).}

\item{nColors}{Number of spectral colors. Default = 5.}

\item{resolution}{The angular resolution. Default = 1.
will generate 90 steps with 1 degree bewteen each step.}
}
\value{
A data.table with one row for each colored ray.
See generateDataMatrix() for details.
}
\description{
Generate the rainbow distributions and plot distributions and maxima.
}
\details{
Will generate a table of traced rays for the selected rainbows and colors.
Rays with uniformly distributed angles between 0 and 90 degrees will
traced through
a water drop the resolution specificies the angular steps.
Each ray will be spectralized with a uniformly distributed light spectrum
of nColors steps bewteen visbleMin and visbleMax. The function will
not perform 3D rendering.
}
\author{
Fredrik Wartenberg
}
