% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRainbows.R
\name{intensities}
\alias{intensities}
\title{Calculate intensities}
\usage{
intensities(pdfData)
}
\arguments{
\item{pdfData}{PDF data (Aggregated raydata)}
}
\value{
data.table with intensities and 3dB width for rainbows
}
\description{
Calculate relative intensities of rainbows
}
\details{
See plotIntensities for details
}
\author{
Fredrik Wartenberg
}
