% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{lambda2f}
\alias{lambda2f}
\title{Wavelength to Frequency}
\usage{
lambda2f(lambda)
}
\arguments{
\item{lambda}{wavelength in nm}
}
\value{
Frequency in HZ
}
\description{
Wavelength to frequency
}
\details{
Uses f = c/lambda
}
\author{
Fredrik Wartenberg
}
