% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRainbows.R
\name{maxima}
\alias{maxima}
\title{Find angels of mximum intensity}
\usage{
maxima(pdfData)
}
\arguments{
\item{pdfData}{data.table with pdf (= aggregated ray data)}
}
\value{
table with maximum & angle for lambda x rainbowNo
}
\description{
Maxima (Peaks) of pdf
}
\details{
find the maxima of the distributions for color x nRainbow
returns a table with maximum & angle for lambda x rainbowNo
Fresnel intensity method is biased towards too small angels.
}
\author{
Fredrik Wartenberg
}
