% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{physicalConstants}
\alias{physicalConstants}
\title{Physical Constants}
\usage{
physicalConstants(print = FALSE)
}
\arguments{
\item{print}{if set to TRUE pretty print the values}
}
\value{
A list of physical constants
}
\description{
Physical constants used in rainbowLab
}
\details{
Inmutable function to print and retrieve constants.
}
\author{
Fredrik Wartenberg
}
