% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotIntensities}
\alias{plotIntensities}
\title{Plot relative rainbow intensities and 3dB width.}
\usage{
plotIntensities(pdfData)
}
\arguments{
\item{pdfData}{data.table with aggregated rayData}
}
\value{
nothing
}
\description{
Plot relative intensities and width of rainbows
}
\details{
Uses intensities calculated by fresnel method.
This is an unverified method. It is
difficult to find actual data on the intensities.
Stated relative intensities between first and second
rainbow range from 50% to 10%.
3dB width is calculated as the difference between the
maximum and the minimum angle where the pdf aggregated
over all lambda is above 50% of the maxiumum value.
(see https://en.wikipedia.org/wiki/Beamwidth)
}
\author{
Fredrik Wartenberg
}
