% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotInVsOut}
\alias{plotInVsOut}
\title{Plot Incident vs. angular difference}
\usage{
plotInVsOut(rayData)
}
\arguments{
\item{rayData}{data.table with rayData}
}
\value{
nothing
}
\description{
Plot angular difference vs incident angle
}
\details{
Mostly a plot to visualize the data. Use plotPDF to
view rainbow angles and distributions.
}
\author{
Fredrik Wartenberg
}
