% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMaxima}
\alias{plotMaxima}
\title{Plot Angles of Maximum Intensities}
\usage{
plotMaxima(pdfData)
}
\arguments{
\item{pdfData}{data.table with aggregated rayData}
}
\value{
data.table with maxima
}
\description{
Plot Angles of Maximum Intensities for Rainbow x Lambda
}
\details{
Use function maxima to calculate data points. See
maxima function for details.
}
\author{
Fredrik Wartenberg
}
