% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPDF}
\alias{plotPDF}
\title{Plot angular difference PDF}
\usage{
plotPDF(pdfData)
}
\arguments{
\item{pdfData}{data.table with aggregated rayData}
}
\value{
nothing
}
\description{
Plot probablity density function for angular difference
}
\details{
Plots the aggregated data in pdfData, no calculation is done here.
For different aggregation options see aggregateData
}
\author{
Fredrik Wartenberg
}
