% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRefractiveIndex}
\alias{plotRefractiveIndex}
\title{Plot Refractive Index}
\usage{
plotRefractiveIndex()
}
\value{
nothing
}
\description{
Plot Refractive Index of water over Lambda
}
\details{
Prints a color coded plot of the refractive index over lambda
Data source: Segelstein 1981; The Complex Refractive Index of Water"
}
\author{
Fredrik Wartenberg
}
