% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRainbows.R
\name{prepareData}
\alias{prepareData}
\title{Prepare rayTracingData for analysis and plotting}
\usage{
prepareData(rayTracingData, simplify = FALSE,
  intensityMethod = "fresnel_m")
}
\arguments{
\item{rayTracingData}{data.table with raytracing data, one row per ray}

\item{simplify}{will remove radian fields if TRUE}

\item{intensityMethod}{Method to calculate intensity, see details}
}
\value{
data.table with added fields, now called rayData
}
\description{
Add fields in degrees and intensity to raytracing data
}
\details{
Intensity is calculated according to intensity method. Currently
supported methods are fresnel (default) and identity
}
\author{
Fredrik Wartenberg
}
