% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ray.R
\name{ray}
\alias{ray}
\title{ray}
\usage{
ray(O = c(0, 0, 0), D = c(1, 0, 0), lambda = 600, alpha = "1")
}
\arguments{
\item{O}{Origin (cartesian coordinates)}

\item{D}{Direction (vector, will be normalised to 1}

\item{lambda}{wavelength [nm]}

\item{alpha}{alpha (for plotting)}
}
\value{
ray object
}
\description{
Generate one ray
}
\details{
A ray represents a light ray. Its properties are
origin (O), direction (D), lambda and alpha.
}
\author{
Fredrik Wartenberg
}
