% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{refractiveIndexFnGenerator}
\alias{refractiveIndexFnGenerator}
\title{generate refractive index function}
\usage{
refractiveIndexFnGenerator(datafile)
}
\arguments{
\item{datafile}{with the measurement values (see details)}
}
\value{
a function to for refractive index
}
\description{
Generator function for refractive index function
}
\details{
This function is used to generate the actual function
for mapping of lambda to refractive index
refractiveIndex(lambda) [in nm]
Source of values:
Segelstein, D., 1981: The Complex Refractive Index of Water,
M.S. Thesis, University of Missouri--Kansas City
http://www.philiplaven.com/Segelstein.txt
refered to from  http://www.philiplaven.com/p20.html
}
\author{
Fredrik Wartenberg
}
