% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{render}
\alias{render}
\title{Dispatch funtction for rendering objects}
\usage{
render(shape, t)
}
\arguments{
\item{shape}{object to render}

\item{t}{unclaer}
}
\value{
nothing
}
\description{
Generic render function
}
\details{
Will dispatch to object's render functions
}
\author{
Fredrik Wartenberg
}
