% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{renderScene}
\alias{renderScene}
\title{Render Scene}
\usage{
renderScene(scene)
}
\arguments{
\item{scene}{scenGraph}
}
\value{
nothing
}
\description{
Render scene
}
\details{
Scene is a the sceneGraph representation of the scene.
render will tranverse the scene graph and call the
render functions of the various objects.
It will do so by calling the generic render function.
rgl is a buggy library, so expect surprises
}
\author{
Fredrik Wartenberg
}
