% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{schlick}
\alias{schlick}
\title{Schlick approximation for reflection}
\usage{
schlick(theta, n = 1.34)
}
\arguments{
\item{theta}{incident angle (radian)}

\item{n}{refractive index of inner medium}
}
\value{
reflection cofficient
}
\description{
Schlick's approximation for the reflection coefficient
}
\details{
Has proved not to work
}
\author{
Fredrik Wartenberg
}
