% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_rainbowLab}
\alias{theme_rainbowLab}
\title{Rainbow Lab Plot theme}
\usage{
theme_rainbowLab()
}
\value{
theme
}
\description{
Theme for rainbow Lab
}
\details{
Based on theme_classic
}
\author{
Fredrik Wartenberg
}
