% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{traceOneRay}
\alias{traceOneRay}
\title{Trace single ray}
\usage{
traceOneRay(nColors = 10, nInteractions = 3, showNormals = FALSE)
}
\arguments{
\item{nColors}{number of colors from the uniform spectrum. Default = 10.}

\item{nInteractions}{interactions of the ray with the drop.
nInteractions -2 gives rainbow number. Default = 3.}

\item{showNormals}{if TRUE, show normals in the interaction point and
plane of interaction. Default = FALSE.}
}
\value{
Nothing
}
\description{
Trace and spectralize one ray and show it in a 3D plot
}
\details{
The drop is located at x = 400 and has a radius of 400
The light source is one single reay
}
\author{
Fredrik Wartenberg
}
