% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFresnel}
\alias{plotFresnel}
\title{Plot Fresenel Coefficients}
\usage{
plotFresnel(n = 1.33, by = 0.1)
}
\arguments{
\item{n}{refractive index}

\item{by}{angular steps (deg) for plotting}
}
\value{
nothing
}
\description{
plot fresnel coefficients for water/air interface
}
\details{
Plots reflection and transmission coeffcients for s and p polarisations
assumes air water interface and n air = 1
}
\author{
Fredrik Wartenberg
}
